Create ValuesGenerated.cpp
The input material parameters are x composition,NA,SRV,Diffusion length,Thickness ,RA.For all the input material parameters except RA ,1,000 samples of normally distributed values are generated with the mean and standard deviation as given in code.The sample space is fairly large for statistical treatment and then further performing the task.One such technique in order to select the input material parameters can be Pareto analysis which is a simple decision-making technique for assessing competing problems and measuring the impact of fixing them. This allows you to focus on solutions that will provide the most benefit.